Add stats page, fix sells process, add receipt printing,etc

+++ database/migrations/2020_11_16_182343_create_items_table.php
@@ -16,7 +16,7 @@ class CreateItemsTable extends Migration
         Schema::create('items', function (Blueprint $table) {
             $table->id();
             $table->string('product_name');
-            $table->decimal('product_price');
+            $table->double('product_price');
             $table->unsignedBigInteger('quantity');
             $table->decimal('sub_total');
             $table->unsignedBigInteger('receipt_id');
